% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.lBF.R
\name{calc.lBF}
\alias{calc.lBF}
\title{calculate BF}
\usage{
calc.lBF(pars = NULL, data.sub = NULL, w0 = 2, log10 = TRUE)
}
\arguments{
\item{pars}{A list object as returned by the function set.parameters}

\item{data.sub}{A 3-dimension array (N. SNPs x N. phenotypes x 3) with state likelihoods.}

\item{w0}{An integer with column index for the null state.}

\item{log10}{Return Bayes Factor in log 10 base.}
}
\value{
lBF Tree Bayes Factor
}
\description{
calculate BF
}
