% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.llk.tree.R
\name{calc.llk.tree}
\alias{calc.llk.tree}
\title{Function to calculate integrated likelihood for set of variants up tree}
\usage{
calc.llk.tree(pars, data.sub, returnForwardMatrix = FALSE,
  returnGMatrix = FALSE)
}
\arguments{
\item{pars}{A list object as returned by the function set.parameters}

\item{data.sub}{A 3-dimension array (N. SNPs x N. phenotypes x 3) with state likelihoods.}

\item{returnForwardMatrix}{Boolean}

\item{returnGMatrix}{Boolean}
}
\value{
Either a numeric object with the Tree likelihood or a list with FWD and G matrices
}
\description{
Function to calculate integrated likelihood for set of variants up tree
}
