% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawTree.R
\name{drawTree}
\alias{drawTree}
\title{Function to draw TreeWAS results}
\usage{
drawTree(tree = NULL, pp = NULL, tree_title = "Tree",
  trim_tree_pp = NULL)
}
\arguments{
\item{tree}{Disease ontology}

\item{pp}{Posterior probabilities table}

\item{tree_title}{Title for plot}

\item{trim_tree_pp}{PP threshold to trim the tree}
}
\value{
Returns a plotly plot with TreeWAS results
}
\description{
Function to draw TreeWAS results
}
