% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{load.HES.lk.data}
\alias{load.HES.lk.data}
\title{Likelihood surfaces for 400 SNPs in the HES analysis of the UK Biobank}
\usage{
load.HES.lk.data()
}
\description{
Likelihood surfaces for 400 SNPs in the HES analysis of the UK Biobank
}
