% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.posterior.profile.R
\name{marginal.posterior.profile}
\alias{marginal.posterior.profile}
\title{Function to get marginal posterior on -/0/+ profile}
\usage{
marginal.posterior.profile(pars = NULL, data.sub = NULL)
}
\arguments{
\item{pars}{A list object as returned by the function set.parameters}

\item{data.sub}{A 3-dimension array (N. SNPs x N. phenotypes x 3) with state likelihoods.}
}
\value{
pp. A 2-dimension array (N. Phenotypes x 3) with posterior probabilties.
}
\description{
Function to get marginal posterior on -/0/+ profile
}
